fix: restructured resume builder
